requirements.txt
torchvision==0.2.1
Keras==2.2.4
h5py==2.7.1
numpy==1.14.0
matplotlib==2.1.2
torch==0.4.1
scipy==1.0.0
Pillow==5.3.0
